Added Lightsabers
Long story short: I found a functional slicing project and copied the important stuff from there. it slices pretty much anything as long as it has a mesh render. current problem is that it doesn't work with skinned mesh renders. Why is this a problem? because cloth components need skinned mesh renderers to work, and the skinned and non-skinned variants are mutually exclusive. mostly. further testing required.