Rewriting of script complete
All positions and rotations of each joint can be found in the assigned arrays.
You have the left and right joint positions and the left and right joint rotations currently.
Each array's entry number corresponds to the ID of each joint.
A list of all joint ID's can be found commented in the "OnHandsUpdate".

From here, we can easily check all of the joint positions and rotations at any given time and do math with them to define various gestures!