Move the world
A small change as requested by Jim
You can now move the world around (which simulates moving the player) by performing the head pat gesture.

Also modified the script that's responsible for this in such a way that you can turn on/off the various things it is responsible for, so you can more easily put the script on other objects.