A good fist
After performing the toolSafeguard, you can now make a fist, which will be recognized as such and then shenanigans. (pinning to be added)

left and right fists kept track of separately

note: the threshold for what counts as a "fist" is pretty lenient. simply holding your fingers somewhat close to your palm is good enough, so long as all of them are close enough at the same time.